requirements.txt
appdirs==1.4.3
apturl==0.5.2
backports.entry-points-selectable==1.1.0
bcrypt==3.1.7
beautifulsoup4==4.9.1
blinker==1.4
Brlapi==0.7.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
cloudinary==1.22.0
colorama==0.4.3
command-not-found==0.3
confusable-homoglyphs==3.2.0
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
distlib==0.3.2
distro==1.4.0
distro-info===0.23ubuntu1
Django==2.2
django-bootstrap4==2.2.0
django-email-verification==0.1.0
django-jazzmin==2.4.8
django-paypal==1.0.0
django-registration==2.4.1
django-tinymce==3.1.0
duplicity==0.8.12.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.0.12
future==0.18.2
getmac==0.8.2
httplib2==0.14.0
idna==2.10
importlib-metadata==1.7.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
monotonic==1.5
more-itertools==4.2.0
netifaces==0.10.4
oauthlib==3.1.0
olefile==0.46
paramiko==2.6.0
pexpect==4.6.0
Pillow==7.2.0
platformdirs==2.0.2
protobuf==3.6.1
psycopg2-binary==2.8.6
pycairo==1.16.2
pycups==1.9.73
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.4
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-decouple==3.3
pytz==2020.1
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.24.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
systemd-python==234
ubuntu-advantage-tools==20.3
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urllib3==1.25.10
usb-creator==0.3.7
vboxapi==1.0
virtualenv==20.0.17
wadllib==1.3.3
whitenoise==5.3.0
xkit==0.0.0
zipp==3.2.0
